dfchronicles/requirements.txt
Django==4.2.4
requests==2.26.0
psycopg==3.1.8
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
openai==0.28.1
python-dotenv==0.19.2
tiktoken==0.5.1
